refactor - move code into individual modules
-Move menus into individual files

move browser windows code to individual files
