+++ app/windows/splashWindow.js
