+++ package.json
@@ -5,7 +5,8 @@
   "main": "main.js",
   "scripts": {
     "start": "node_modules/.bin/electron .",
-    "test": "node_modules/.bin/mocha test/front.test.js"
+    "test": "node_modules/.bin/mocha test/front.test.js",
+    "test-with-coverage": "nyc --reporter=text mocha"
   },
   "repository": {
     "type": "git",
@@ -24,6 +25,7 @@
   "devDependencies": {
     "chai": "^4.1.2",
     "electron": "^2.0.8",
+    "istanbul": "^0.4.5",
     "mocha": "^5.2.0",
     "spectron": "^3.8.0"
   }
